+++ src/main/python/apache/aurora/client/cli/task.py
@@ -17,7 +17,11 @@
 
 from __future__ import print_function
 
+import os
 import subprocess
+from argparse import ArgumentTypeError
+
+from pystachio import Environment, String
 
 from apache.aurora.client.api.command_runner import (
     DistributedCommandRunner,
@@ -30,12 +34,17 @@ from apache.aurora.client.cli.options import (
     ALL_INSTANCES,
     EXECUTOR_SANDBOX_OPTION,
     INSTANCES_SPEC_ARGUMENT,
+    SCP_DEST_ARGUMENT,
+    SCP_OPTIONS,
+    SCP_SOURCE_ARGUMENT,
     SSH_INSTANCE_ARGUMENT,
     SSH_OPTIONS,
     SSH_USER_OPTION,
-    CommandOption
+    CommandOption,
+    parse_task_instance_key
 )
 from apache.aurora.common.clusters import CLUSTERS
+from apache.thermos.config.schema import ThermosContext
 
 
 class RunCommand(Verb):
@@ -148,6 +157,91 @@ class SshCommand(Verb):
     return subprocess.call(ssh_command)
 
 
+class ScpCommand(Verb):
+
+  JOB_NOT_FOUND_ERROR_MSG = 'Job or instance %s/%s not found'
+  TILDE_USAGE_ERROR_MSG = 'Command does not support tilde expansion for path: %s'
+
+  @property
+  def name(self):
+    return 'scp'
+
+  @property
+  def help(self):
+    return """executes an scp to/from/between task instance(s). The task sandbox acts as
+  the relative root.
+  """
+
+  @staticmethod
+  def _extract_task_instance_and_path(context, file_path):
+    key = file_path.split(':', 1)
+    try:
+      if (len(key) == 1):
+        return (None, key[0])  # No jobkey specified
+      return (parse_task_instance_key(key[0]), key[1])
+    except ArgumentTypeError as e:
+      raise context.CommandError(EXIT_INVALID_PARAMETER, str(e))
+
+  @staticmethod
+  def _build_path(context, target):
+    (task_instance, path) = ScpCommand._extract_task_instance_and_path(context, target)
+
+    # No jobkey is specified therefore we are using a local path.
+    if (task_instance is None):
+      return path
+
+    # Jobkey specified, we want to convert to the user@host:file scp format
+    (cluster, role, env, name) = task_instance.jobkey
+    instance = set([task_instance.instance])
+    api = context.get_api(cluster)
+    resp = api.query(api.build_query(role, name, env=env, instances=instance))
+    context.log_response_and_raise(resp,
+        err_msg=('Unable to get information about instance: %s' % combine_messages(resp)))
+    if (resp.result.scheduleStatusResult.tasks is None or
+        len(resp.result.scheduleStatusResult.tasks) == 0):
+      raise context.CommandError(EXIT_INVALID_PARAMETER,
+          ScpCommand.JOB_NOT_FOUND_ERROR_MSG % (task_instance.jobkey, task_instance.instance))
+    first_task = resp.result.scheduleStatusResult.tasks[0]
+    assigned = first_task.assignedTask
+    role = assigned.task.job.role
+    slave_host = assigned.slaveHost
+
+    # If path is absolute, use that. Else if it is a tilde expansion, throw an error.
+    # Otherwise, use sandbox as relative root.
+    normalized_input_path = os.path.normpath(path)
+    if (os.path.isabs(normalized_input_path)):
+      final_path = normalized_input_path
+    elif (normalized_input_path.startswith('~/') or normalized_input_path == '~'):
+      raise context.CommandError(EXIT_INVALID_PARAMETER, ScpCommand.TILDE_USAGE_ERROR_MSG % path)
+    else:
+      sandbox_path_pre_format = DistributedCommandRunner.thermos_sandbox(
+          api.cluster,
+          executor_sandbox=context.options.executor_sandbox)
+      thermos_namespace = ThermosContext(
+          task_id=assigned.taskId,
+          ports=assigned.assignedPorts)
+      sandbox_path = String(sandbox_path_pre_format) % Environment(thermos=thermos_namespace)
+      # Join the individual folders to the sandbox path to build safely
+      final_path = os.path.join(str(sandbox_path), *normalized_input_path.split(os.sep))
+
+    return '%s@%s:%s' % (role, slave_host, final_path)
+
+  def get_options(self):
+    return [
+        SCP_SOURCE_ARGUMENT,
+        SCP_DEST_ARGUMENT,
+        SCP_OPTIONS,
+        EXECUTOR_SANDBOX_OPTION
+    ]
+
+  def execute(self, context):
+    scp_command = ['scp']
+    scp_command += context.options.scp_options if context.options.scp_options else []
+    scp_command += [ScpCommand._build_path(context, p) for p in context.options.source]
+    scp_command += [ScpCommand._build_path(context, context.options.dest)]
+    return subprocess.call(scp_command)
+
+
 class Task(Noun):
   @property
   def name(self):
@@ -165,3 +259,4 @@ class Task(Noun):
     super(Task, self).__init__()
     self.register_verb(RunCommand())
     self.register_verb(SshCommand())
+    self.register_verb(ScpCommand())